# Conflicts:
#	src/main/java/DoAnChuyenNganh/WebsiteChamSocThuCung/controllers/AppointmentController.java
#	src/main/java/DoAnChuyenNganh/WebsiteChamSocThuCung/repositories/AppointmentRepository.java
#	src/main/java/DoAnChuyenNganh/WebsiteChamSocThuCung/services/AppointmentService.java
#	src/main/resources/templates/appointments/appointment-list.html
#	src/main/resources/templates/appointments/create-appointment.html

+++ src/main/java/DoAnChuyenNganh/WebsiteChamSocThuCung/models/Category.java
@@ -17,5 +17,5 @@ public class Category {
     @NotBlank(message = "Tên là bắt buộc")
     private String name;
     @Column(name = "IS_REMOVED")
-    private byte isRemoved;
+    private int isRemoved;
 }
